updated js files, accessed router
