+++ views/layout.ejs
@@ -5,9 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Pokedex</title>
     <!-- Materialize CSS -->
-    <link rel="stylesheet" href="public/css/style.css">
-
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0-beta/css/materialize.min.css">
+    <link rel="stylesheet" type="text/css" href="/css/style.css">
   </head>
   <body>
         <nav>
@@ -17,7 +16,6 @@
                 PokeDex
               </a>
               <ul id="nav-mobile" class="right hide-on-med-and-down">
-                <li><a class="yellow-text text-accent-3" href="/">Home</a></li>
                 <li><a class="yellow-text text-accent-3" href="/">Pokemon!</a></li>
                 <li><a class="yellow-text text-accent-3" href="/pokemon">My Favorites</a></li>
               </ul>
@@ -26,7 +24,8 @@
     <%- body %>
     <!-- jQuery -->
     <script src="//code.jquery.com/jquery.js"></script>
-    <!-- Bootstrap JavaScript -->
+    <!-- Materialize JavaScript -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0-beta/js/materialize.min.js"></script>
+    <script src="/js/script.js"></script>
   </body>
 </html>
